tuned webpack eslint package.json and split code on modules
