+++ .eslintrc.yml
@@ -1,6 +1,13 @@
+parserOptions:
+  ecmaVersion: 2015
+  sourceType: 'module'
+
 env:
   es6: true
   browser: true
   commonjs: true
 
 extends: 'htmlacademy/es6'
+
+rules:
+  object-curly-spacing: ["error", "always"]
