+++ public/index.html
@@ -36,7 +36,7 @@
       <p>130 291 movies inside</p>
     </section>
   </footer>
-  <script src="../src/main.js"></script>
+  <script src="./bundle.js"></script>
 </body>
 
 </html>
