+++ src/components/show-more-button.js
@@ -1,0 +1,3 @@
+const createShowMoreButtonTemplate = () => `<button class="films-list__show-more">Show more</button>`;
+
+export { createShowMoreButtonTemplate };
